create a WelcomeWidgets
