+++ lib/core/routes/routes.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:my_market/feature/auth/logic/cubit/auth_cubit.dart';
-import 'package:my_market/feature/auth/views/signin_view.dart';
+import 'package:my_market/feature/auth/views/screens/signin_view.dart';
 
-import '../../feature/auth/views/login_view.dart';
+import '../../feature/auth/views/screens/login_view.dart';
 import 'app_routes.dart';
 
 class PageRoutes extends MaterialPageRoute {
